+++ node/resolvers/search/index.ts
@@ -27,6 +27,9 @@ import { resolvers as productSearchResolvers } from './productSearch'
 import { resolvers as recommendationResolvers } from './recommendation'
 import { resolvers as breadcrumbResolvers } from './searchBreadcrumb'
 import { resolvers as skuResolvers } from './sku'
+import { resolvers as skuSpecificationResolver } from './skuSpecification'
+import { resolvers as skuSpecificationFieldResolver } from './skuSpecificationField'
+import { resolvers as skuSpecificationValueResolver } from './skuSpecificationValue'
 import { resolvers as productPriceRangeResolvers } from './productPriceRange'
 import {
   SearchCrossSellingTypes,
@@ -138,6 +141,9 @@ export const fieldResolvers = {
   ...productResolvers,
   ...recommendationResolvers,
   ...skuResolvers,
+  ...skuSpecificationResolver,
+  ...skuSpecificationFieldResolver,
+  ...skuSpecificationValueResolver,
   ...breadcrumbResolvers,
   ...productSearchResolvers,
   ...assemblyOptionResolvers,