+++ node/resolvers/search/skuSpecificationField.ts
@@ -1,0 +1,15 @@
+import { addContextToTranslatableString } from '../../utils/i18n'
+
+export const resolvers = {
+  SKUSpecificationField: {
+    name: ( field: SKUSpecificationField, _: any, ctx: Context) => {
+      return addContextToTranslatableString(
+        {
+          content: field.name,
+          context: field.id
+        },
+        ctx
+      )
+    }
+  }
+}