+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Translation to `SKUSpecificationField` and `SKUSpecificationValue`.
 
 ## [0.8.1] - 2020-07-07
 
@@ -13,7 +15,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Slugify facet keys before check if it is selected.
 
 ## [0.8.0] - 2020-06-17
-
 ### Added
 - Add `suggestions`, `correction`, `banners`, and `redirect`.